+++ projects/np-ui-lib/src/lib/np-accordion/np-accordion-item/np-accordion-item.component.ts
@@ -83,4 +83,8 @@ export class NpAccordionItemComponent implements OnInit, OnDestroy {
     }
   }
 
+  _getTitleId() {
+    return this.inputId + '_title';
+  }
+
 }