+++ projects/np-ui-lib/src/lib/np-color-picker/np-color-picker.component.ts
@@ -52,6 +52,7 @@ export class NpColorPickerComponent implements ControlValueAccessor, AfterViewIn
   isShowCursorDiv = false;
   innerValue: string;
   isDisabled = false;
+  focused = false;
   private templatePortal: TemplatePortal<any>;
   private overlayRef: OverlayRef;
   private onChangeCallback: (_: any) => void = () => { };
@@ -325,7 +326,12 @@ export class NpColorPickerComponent implements ControlValueAccessor, AfterViewIn
     }
   }
 
+  _onFocus() {
+    this.focused = true;
+  }
+
   _onBlur() {
+    this.focused = false;
     this.onTouchedCallback();
   }
 }