+++ projects/np-ui-lib/src/lib/np-date-picker/np-date-picker.component.ts
@@ -61,6 +61,7 @@ export class NpDatePickerComponent implements ControlValueAccessor, AfterViewIni
   isOpen = false;
   innerValue: Date;
   isDisabled = false;
+  focused = false;
 
   private templatePortal: TemplatePortal<any>;
   private overlayRef: OverlayRef;
@@ -438,7 +439,12 @@ export class NpDatePickerComponent implements ControlValueAccessor, AfterViewIni
     this._calculateDays();
   }
 
+  _onFocus() {
+    this.focused = true;
+  }
+
   _onBlur() {
+    this.focused = false;
     this.onTouchedCallback();
   }
 }