+++ projects/np-ui-lib/src/lib/np-loader/np-loader.directive.ts
@@ -18,6 +18,8 @@ export class NpLoaderDirective implements OnChanges, OnInit {
 
     ngOnInit(): void {
         const loader = this.renderer.createElement('div');
+        this.renderer.setAttribute(loader, 'role', 'alert');
+        this.renderer.setAttribute(loader, 'aria-busy', 'true');
         this.renderer.addClass(loader, 'np-loader');
         this.renderer.setStyle(loader, 'width', `${this.diameter}px`);
         this.renderer.setStyle(loader, 'height', `${this.diameter}px`);