+++ projects/np-ui-lib/src/lib/np-switch/np-switch.component.ts
@@ -30,6 +30,7 @@ export class NpSwitchComponent implements ControlValueAccessor {
 
   innerValue: boolean;
   isDisabled = false;
+  focused = false;
   private onChangeCallback: (_: any) => void = () => { };
   private onTouchedCallback: () => void = () => { };
 
@@ -71,7 +72,12 @@ export class NpSwitchComponent implements ControlValueAccessor {
     this.value = $event.target.checked;
   }
 
+  _onFocus() {
+    this.focused = true;
+  }
+
   _onBlur() {
+    this.focused = false;
     this.onTouchedCallback();
   }
 }