+++ projects/np-ui-lib/src/lib/np-time-picker/np-time-picker.component.ts
@@ -48,6 +48,7 @@ export class NpTimePickerComponent implements ControlValueAccessor, AfterViewIni
   pattern: any;
   innerValue: string;
   isDisabled = false;
+  focused = false;
   private templatePortal: TemplatePortal<any>;
   private overlayRef: OverlayRef;
   private onChangeCallback: (_: any) => void = () => { };
@@ -379,7 +380,12 @@ export class NpTimePickerComponent implements ControlValueAccessor, AfterViewIni
     this._extractValues();
   }
 
+  _onFocus() {
+    this.focused = true;
+  }
+
   _onBlur() {
+    this.focused = false;
     this.onTouchedCallback();
   }
 }